Fixed #34 example.
I'm sorely tempted to revert the order of the arguments again and
place the rel at the start - mostly due to the more formalization
that a link/resource has a _single_ reltype and not, as previously
implied a space separated list of reltypes.

But that can wait for 2.0 - breaking change.